fix problem

+++ src/wmw24/sasasightsee/client/OSMObject.java
@@ -1,28 +1,29 @@
+
 package wmw24.sasasightsee.client;
 
 import com.google.gwt.core.client.JavaScriptObject;
 
 public class OSMObject extends JavaScriptObject
 {
-	protected OSMObject()
-	{
+   protected OSMObject()
+   {
 
-	}
+   }
 
-	public final native String getType() /*-{
+   public final native String getType() /*-{
 		return this.type;
-	}-*/;
+   }-*/;
 
-	public final native long getId() /*-{
+   public final native int getId() /*-{
 		return this.id;
-	}-*/;
+   }-*/;
 
-	public final native double getLat() /*-{
+   public final native double getLat() /*-{
 		return this.lat;
-	}-*/;
+   }-*/;
 
-	public final native double getLon() /*-{
+   public final native double getLon() /*-{
 		return this.lon;
-	}-*/;
+   }-*/;
 
 }
